Refactor hivemind.Server
